+++ Changes.md
@@ -6,6 +6,11 @@ API
 
 - Signals : Removed usage of `boost::signals::detail::unusable` as a substitute for the `void` return type in the Signal bindings. Custom SlotCallers may now use a standard `void` return type.
 
+Breaking Changes
+----------------
+
+- Replaced all usage of `boost::optional` with `std::optional`.
+
 Build
 -----